// mantenedorsmart/src/app/models/periodo.ts
export interface PeriodoType {
    idControlPeriodo:    string;
    numeroPeriodo:       number;
    anioFiscal:          string;
    inicioProcesamiento: Date;
    finProcesamiento:    Date;
    isActivo:            boolean;
    periodoId:           number;
    nombrePeriodo:       string;
    anioId:              number;
}